[Loom] Isolate bytecode catalog numbering
Separate stable symbol order, first-use string and operation IDs,
and structural type deduplication from file framing. Attribute and
encoding dependency numbering remains coupled to the catalog so every
wire table is complete before section emission.
